Allowing for an HTTP server to control processor parameters
Initial implementation of a RESTful HTTP API for controlling processor
parameters. The idea is that OpenEphys GUI can be a part of a larger
ecosystem during a running experiment, and these components might need
to communicate with one another. Adding an HTTP API such as one here
allows the setting / getting of parameters to synchronize different
components (e.g. filter settings, session names, etc.).

This starts an HTTP server on port 37497 (== "EPHYS" on a phone keypad).