Supporting N-D PCA instead of hardcoding 2
Relaxes the underlying infrastructure to support more than 2 PCs, though
the (a) matching process with a polygon is still 2D, and (b) the display
is fixed to the first 2 components.

Also it seems that the loading/saving of the PCA details into the XML
was broken before this. The saving of the PCA details would save under
the <ELECTRODE> node, but the loading was looking for the PCA details in
the <SPIKESORTING> node.