Allowing handling, setting, and drawing of N-D ellipses
Allowing the use of N-D ellipses, in addition to 2-D hand-drawn
polygons, for spike sorting. These changes include adapting PointD to be
fully N-dimensional (instead of a hard-coded 2-dimensions), grouping the
PCA details together and allowing serialization/deserialization via
JSON/the HTTP API, and the display of the ellipse in 2 dimensions.