Correctly enabling/disabling all editors in the chain
Previous behavior did not correctly enable all the editors in the chain,
respecting splitters & mergers. Instead, it would only enable one "side"
of the splitter, and thus attempts to start the recording session would
fail, unless you literally clicked on all of the paths of all of the
splitters.

This does a depth-first traversal of the editors and propagates the
enabled bit down the chain as necessary.